+++ Python/models/dia.py
@@ -1,5 +1,4 @@
-from sqlalchemy import *
-from sqlalchemy import UniqueConstraint
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, Numeric, String, UniqueConstraint
 from database import Base
 from sqlalchemy.orm import relationship
 
@@ -10,7 +9,7 @@ class Dia(Base):
     client_id = Column(String(50), index=True, nullable=True)
     ciclo_id = Column(Integer, ForeignKey("ciclos.id"))
     data = Column(DateTime(timezone=True))
-    saldo = Column(Float)
+    saldo = Column(Numeric(10, 2))
 
     ciclo = relationship("Ciclo", back_populates="dias")
     gastos = relationship("Gasto", back_populates="dia", cascade="all, delete-orphan")
